pyproject.toml
[project]
name = "event-service"
version = "0.1.0"
description = "An API for admin of sporting events"
readme = "README.md"
requires-python = ">=3.11.9"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
authors = [
    { name = "Stig B. Dørmænen", email = "stigbd@gmail.com" },
    { name = "Heming Harnæs", email = "heming.harnaes@gmail.com" },
]
dependencies = [
    "PyJWT>=2.1.0",
    "aiodns>=3.0.0",
    "aiofiles>=25.1.0",
    "certifi>=2024.12.14",
    "fastapi[standard]>=0.128.1",
    "httpx>=0.28.1",
    "motor>=3.3.2",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pydantic>=2.12.5",
]
[project.urls]
Homepage = "https://github.com/langrenn-sprint/event-service"
Repository = "https://github.com/langrenn-sprint/event-service"

[dependency-groups]
dev = [
    "Pygments>=2.10.0",
    "codecov>=2.1.10",
    "coverage>=7.1.0",
    "docker>=7.0.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-docker>=3.1.1",
    "pytest-mock>=3.8.2",
    "pip-audit>=2.7.3",
    "ruff>=0.8.4",
    "poethepoet>=0.32.0",
    "deptry>=0.21.2",
    "pytest-env>=1.1.5",
    "ty>=0.0.12",
    "httpx>=0.28.1",
]

[tool.uv]
required-version = ">=0.10.2"

[tool.ruff.lint]
select = ["ALL"]
# and then manually ignore annoying ones:
ignore = ["COM812", "ISC001", "T201", "G004", "E501", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["ERA001"]
"tests/**/*.py" = [
     "S101",     # asserts allowed in tests...
     "ANN201",   # Missing return type annotation -> tests are often more readable without return type annotations...
     "ARG001",   # Unused function args -> fixtures nevertheless are functionally relevant...
     "ERA001",   # Don't care about commented-out code in tests
     "PLR0913",  # Don't care about too many arguments in test
     "PLR0915",  # Don't care about too many statements in tests
     "PLR2004",  # Don't care about magic value in tests
]

[tool.ruff.lint.isort]
# so it knows to group first-party stuff last
known-first-party = ["app"]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.deptry]
ignore_notebooks = true
known_first_party = ["app"]

[tool.deptry.per_rule_ignores]
DEP002 = ["aiodns", "certifi", "PyJWT"]

[tool.coverage.paths]
source = ["app", "*/site-packages"]

[tool.coverage.run]
branch = false
omit = ["*gunicorn_config.py", "*db_utils.py"]
source = ["app"]

[tool.coverage.report]
fail_under = 100
show_missing = true


[tool.pytest.ini_options]
env = [
    "CONFIG=test",
    "JWT_SECRET=test_jwt_secret_key_for_development_do_not_use_in_production_12345",
    "ADMIN_USERNAME=admin",
    "ADMIN_PASSWORD=password",
    "JWT_EXP_DELTA_SECONDS=60",
    "DB_NAME=users_test",
    "DB_USER=admin",
    "DB_PASSWORD=admin",
    "LOGGING_LEVEL=INFO",
    "EVENTS_HOST_SERVER=localhost",
    "EVENTS_HOST_PORT=8081",
    "COMPETITION_FORMAT_HOST_SERVER=localhost",
    "COMPETITION_FORMAT_HOST_PORT=8082",
    "USERS_HOST_SERVER=localhost",
    "USERS_HOST_PORT=8081",
]

asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit (fast)",
    "integration: marks tests as integration (slower)",
    "contract: marks test as contract (slow)",
]

[tool.poe.tasks]
format = { cmd = "uv run ruff format" }
lint = { cmd = "uv run ruff check --fix" }
check-types = "uv run ty check ."
check-deps = { cmd = "uv run deptry ." }
audit = { cmd = "uv run pip-audit" }
unit-test = { cmd = "uv run pytest -m unit", env = { "CONFIG" = "test" } }
integration-test = { cmd = "uv run pytest -m integration -s --cov --cov-report=term-missing --cov-report=html:.htmlcov", env = { "CONFIG" = "test" } }
contract-test = { cmd = "uv run pytest -m contract -s -rA" }
release = { sequence = [
    "lint",
    "check-types",
    "check-deps",
    "audit",
    "unit-test",
    "integration-test",
    "contract-test",
] }
